requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
astroid==2.15.0
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
exceptiongroup==1.1.0
execnet==1.9.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jproperties==2.1.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.0
platformdirs==3.1.0
pluggy==1.0.0
prettyconf==2.2.1
pycodestyle==2.10.0
pydantic==1.10.6
pylint==2.16.4
pyrsistent==0.19.3
pytest==7.2.1
pytest-rerunfailures==11.1.2
pytest-retry==1.2.1
pytest-xdist==3.1.0
requests==2.28.2
requests-to-curl==1.1.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.15.0
xmltodict==0.11.0